- Unit test for delete service in Open API. - Added a validation for user access on hierarchy before deleting a folder.

+++ app/src/com/constellio/app/modules/restapi/folder/FolderService.java
@@ -130,6 +130,7 @@ public class FolderService extends ResourceService {
 		Record folder = getRecord(id, false);
 		User user = getUser(serviceKey, folder.getCollection());
 		validateUserAccess(user, folder, method);
+		validateUserDeleteAccessOnHierarchy(user, folder);
 
 		folderDao.deleteFolder(user, folder, Boolean.TRUE.equals(physical));
 	}
